+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.5.0] - 2021-08-15
+
 ### Added
 
 -   Support `--force` in interactive mode to allow users to repatch after a game update.
@@ -61,7 +63,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Initial release.
 
-[Unreleased]: https://github.com/nbusseneau/hephaistos/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/nbusseneau/hephaistos/compare/v0.5.0...HEAD
+
+[v0.5.0]: https://github.com/nbusseneau/hephaistos/compare/v0.4.0...v0.5.0
 
 [v0.4.0]: https://github.com/nbusseneau/hephaistos/compare/v0.3.0...v0.4.0